requirements.txt
Flask
Werkzeug
pillow
pytesseract
pymysql

